Implement resource authorization for job deletion

+++ App/controllers/ListingController.php
@@ -5,6 +5,8 @@ namespace App\Controllers;
 use Error;
 use Framework\Database;
 use Framework\Validation;
+use Framework\Session;
+use Framework\Authorization;
 use PDO;
 
 
@@ -22,7 +24,7 @@ class ListingController
     public function index()
     {
 
-        $listings = $this->db->query("SELECT * FROM listings ")->fetchAll(PDO::FETCH_OBJ);
+        $listings = $this->db->query("SELECT * FROM listings ORDER BY created_at DESC")->fetchAll(PDO::FETCH_OBJ);
 
 
         loadView('listings/index', ['listings' => $listings]);
@@ -60,7 +62,7 @@ class ListingController
 
         $newListingData = array_intersect_key($_POST, array_flip($allowedFields));
 
-        $newListingData['user_id'] = 1;
+        $newListingData['user_id'] = Session::get('user')['id'];
 
         $newListingData = array_map('sanitize', $newListingData);
 
@@ -118,10 +120,19 @@ class ListingController
 
         $listing = $this->db->query("SELECT * FROM listings WHERE id = :id", $params)->fetch();
 
+        //Check if listing exists
         if (!$listing) {
             ErrorController::notFound('Listing not found');
             return;
         }
+
+        //Authorization
+        if (!Authorization::isOwner($listing->user_id)) {
+            $_SESSION['error_message'] = 'You are not authorized to delete this listing';
+            return redirect("/listings/{$listing->id}");
+        }
+
+
 
         $this->db->query("DELETE FROM listings WHERE id = :id", $params);
 
@@ -148,13 +159,13 @@ class ListingController
         loadView('listings/edit', ['listing' => $listing]);
     }
 
-         /*/**
+    /*/**
          * Update listing
          * 
          * @param array $params
          * return variant
          */
-        
+
 
     public function update($params)
     {
